requirements.txt
aniso8601==1.3.0
blinker==1.4
click==6.7
Flask==0.12.2
Flask-Breadcrumbs==0.4.0
Flask-Mail==0.9.1
Flask-Menu==0.7.0
Flask-MySQL==1.4.0
Flask-MySQLdb==0.2.0
Flask-RESTful==0.3.6
Flask-Stache==0.1.1
Flask-WTF==0.14.2
gunicorn==19.7.1
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
mysqlclient==1.3.12
numpy==1.13.3
pandas==0.21.1
passlib==1.7.1
PyMySQL==0.7.11
pystache==0.5.4
python-dateutil==2.6.1
pytz==2017.3
six==1.11.0
SQLAlchemy==1.1.15
Werkzeug==0.13
WTForms==2.1
